Add files via upload
A measurement of the matter variation in the local Universe and of $\sigma_8$ is obtained through neutral atomic hydrogen (HI) sources provided by the ALFALFA catalog. By examining the observational data of nearby galaxies, we aim to quantify the spatial correlations between these objects and thus determine the amplitude and correlation scale of density fluctuations in the local Universe, which allowed us to find 
$\sigma_8 = 0.78 \pm 0.04$  for $h =  0.6727$, $\sigma_8 = 0.80 \pm 0.05$ for $h = 0.6980 $ and $\sigma_8 = 0.83 \pm 0.05$ for $h = 0.7304$. A value of $H_0$ is necessary to compare our primary measurements, where the scale is in units of Mpc, with measurements expressed in units of $h^{-1}$ Mpc.

The correlation function $\xi(r)$ provides a systematic approach to analyze these correlations, allowing us to investigate how the properties of the matter distribution vary on different spatial scales. Moreover, the comparison between observational data and theoretical models allows us to test the validity of the standard cosmological model and refine our understanding of the underlying physics.